Make setNamespaceLayoutContext to be responsible only for setting context
