+++ src/lib/layout/api-namespace.ts
@@ -1,5 +1,4 @@
-import type { Blueprint, Route } from '@seamapi/blueprint'
-import type Metalsmith from 'metalsmith'
+import type { Blueprint } from '@seamapi/blueprint'
 
 import type { PathMetadata } from 'lib/path-metadata.js'
 
@@ -9,52 +8,33 @@ export interface ApiNamespaceLayoutContext {
   resources: Array<{ name: string; description: string }>
 }
 
-type File = ApiNamespaceLayoutContext & { layout: string }
-
 export function setNamespaceLayoutContext(
-  files: Metalsmith.Files,
-  blueprint: Blueprint,
+  file: ApiNamespaceLayoutContext,
+  namespace: string,
+  resources: Blueprint['resources'],
   pathMetadata: PathMetadata,
 ): void {
-  const namespacePaths = getNamespacePaths(blueprint.routes)
-
-  for (const path of namespacePaths) {
-    const namespaceMetadata = pathMetadata[path]
-    if (namespaceMetadata == null) continue
-
-    const k = `api${path}/README.md`
-    files[k] = {
-      contents: Buffer.from('\n'),
-    }
-    const file = files[k] as unknown as File
-    file.layout = 'api-namespace.hbs'
-
-    file.title = namespaceMetadata.title
-    file.description = namespaceMetadata.description ?? ''
-    const namespaceResources = Object.entries(pathMetadata)
-      .filter(([p]) => p.startsWith(path))
-      .flatMap(([_, metadata]) => metadata.resources)
-    file.resources = namespaceResources.map((resourceName) => {
-      const resource = blueprint.resources[resourceName]
-
-      if (resource == null) {
-        throw new Error(`Resource ${resourceName} not found in blueprint`)
-      }
-
-      return {
-        name: resourceName,
-        description: resource.description,
-      }
-    })
+  const namespaceMetadata = pathMetadata[namespace]
+  if (namespaceMetadata == null) {
+    throw new Error(`Namespace metadata for ${namespace} not found`)
   }
-}
 
-function getNamespacePaths(routes: Route[]): string[] {
-  return Array.from(
-    new Set(
-      routes.flatMap((route) =>
-        route.namespace != null ? [route.namespace.path] : [],
-      ),
-    ),
-  )
+  file.title = namespaceMetadata.title
+  file.description = namespaceMetadata.description ?? ''
+
+  const namespaceResources = Object.entries(pathMetadata)
+    .filter(([p]) => p.startsWith(namespace))
+    .flatMap(([_, metadata]) => metadata.resources)
+  file.resources = namespaceResources.map((resourceName) => {
+    const resource = resources[resourceName]
+
+    if (resource == null) {
+      throw new Error(`Resource ${resourceName} not found in blueprint`)
+    }
+
+    return {
+      name: resourceName,
+      description: resource.description,
+    }
+  })
 }
